Refactor: bigrams out of scorer
* Take bigram related code out of scorer/ folder
* Fix mypy errors
* Fix ruff errors (all but TODOs)
