+++ granite_tools/bigram_scores/__init__.py
@@ -1,0 +1,3 @@
+from .bigram_scores import (
+    load_bigram_and_unigram_scores as load_bigram_and_unigram_scores,
+)
