+++ granite_tools/scorer/__init__.py
@@ -1,6 +1,3 @@
-from .bigram_scores import (
-    load_bigram_and_unigram_scores as load_bigram_and_unigram_scores,
-)
 from .data import get_trigram_data as get_trigram_data
 from .data import get_trigram_data_from_files as get_trigram_data_from_files
 from .scorer import TrigramModelParameters as TrigramModelParameters
