+++ granite_tools/scripts/check_sentence_score.py
@@ -15,9 +15,9 @@ import sys
 import numpy as np
 import pandas as pd
 
+from granite_tools.bigram_scores import load_bigram_and_unigram_scores
 from granite_tools.config import read_config
 from granite_tools.hands import get_hands_data
-from granite_tools.scorer.bigram_scores import load_bigram_and_unigram_scores
 from granite_tools.scorer.scorer import TrigramModelParameters, get_score
 
 pd.set_option("display.max_rows", 1000)
