+++ granite_tools/scripts/scoreratios_plot.py
@@ -16,15 +16,15 @@ import typing
 
 from matplotlib import pyplot as plt
 
-from granite_tools.config import read_config
-from granite_tools.hands import get_hands_data
-from granite_tools.scorer.bigram_scores import load_ranking
-from granite_tools.scorer.modelparams import SPLINE_KWARGS
-from granite_tools.scorer.smooth_scores import (
+from granite_tools.bigram_scores.anchor_scores import read_raw_anchor_scores_json
+from granite_tools.bigram_scores.rankings import load_bigram_rankings
+from granite_tools.bigram_scores.spline_smoothing import (
+    SPLINE_KWARGS,
     create_monotone_bspline,
-    read_raw_anchor_scores_json,
     scores_to_training_data,
 )
+from granite_tools.config import read_config
+from granite_tools.hands import get_hands_data
 
 if typing.TYPE_CHECKING:
     from typing import Sequence
@@ -56,7 +56,7 @@ if __name__ == "__main__":
 
     config = read_config(config_file)
     hands = get_hands_data(config)
-    ngrams_ordered = load_ranking(bigram_ranking_file)
+    ngrams_ordered = load_bigram_rankings(bigram_ranking_file)
     scores = read_raw_anchor_scores_json(anchor_scores_raw_json_file)
     x_train, y_train, x_all = scores_to_training_data(ngrams_ordered, scores)
