+++ pyproject.toml
@@ -32,7 +32,7 @@ granite-bigram-ranking-initial = "granite_tools.bigram_ranking_initial.bigram_ra
 granite-bigram-ranking-view = "granite_tools.viewer.viewer_app:main"
 granite-bigram-compare = "granite_tools.bigram_compare.compare_app:main"
 granite-bigram-compare-merge = "granite_tools.bigram_compare.merge_data:main"
-granite-bigram-score-ratio-template = "granite_tools.scorer.cli:create_bigram_score_ratio_template"
+granite-bigram-score-ratio-template = "granite_tools.bigram_scores.cli:create_bigram_score_ratio_template"
 granite-trigram-score-template = "granite_tools.scorer.cli:create_score_template_cli"
 granite-trigram-score-fit = "granite_tools.scorer.cli:fit_parameters_cli"
 granite-trigram-score-fit-check = "granite_tools.scorer.cli:fit_check_cli"
