+++ tests/test_bigram_compare/test_compare_app.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import pickle
 from pathlib import Path
